Update example mods

+++ mods/GameSession-KillStats/GameSession.lua
@@ -18,7 +18,7 @@ GameSession.Event = {
 	Update = 'Update',
 	Load = 'Load',
 	Save = 'Save',
-	List = 'List',
+	Clean = 'Clean',
 	LoadData = 'LoadData',
 	SaveData = 'SaveData',
 }
@@ -39,7 +39,7 @@ local eventScopes = {
 	[GameSession.Event.Update] = {},
 	[GameSession.Event.Load] = { [GameSession.Scope.Saves] = true },
 	[GameSession.Event.Save] = { [GameSession.Scope.Saves] = true },
-	[GameSession.Event.List] = { [GameSession.Scope.Saves] = true },
+	[GameSession.Event.Clean] = { [GameSession.Scope.Saves] = true },
 	[GameSession.Event.LoadData] = { [GameSession.Scope.Saves] = true, [GameSession.Scope.Persistence] = true },
 	[GameSession.Event.SaveData] = { [GameSession.Scope.Saves] = true, [GameSession.Scope.Persistence] = true },
 }
@@ -291,6 +291,10 @@ local function importSession(s)
 end
 
 local function writeSession(sessionName, sessionData)
+	if not sessionDataDir then
+		return
+	end
+
 	local sessionPath = sessionDataDir .. '/' .. sessionName .. '.lua'
 	local sessionFile = io.open(sessionPath, 'w')
 
@@ -304,6 +308,10 @@ local function writeSession(sessionName, sessionData)
 end
 
 local function readSession(sessionName)
+	if not sessionDataDir then
+		return nil
+	end
+
 	local sessionPath = sessionDataDir .. '/' .. sessionName .. '.lua'
 	local sessionChunk = loadfile(sessionPath)
 
@@ -320,12 +328,20 @@ local function readSession(sessionName)
 end
 
 local function removeSession(sessionName)
+	if not sessionDataDir then
+		return
+	end
+
 	local sessionPath = sessionDataDir .. '/' .. sessionName .. '.lua'
 
 	os.remove(sessionPath)
 end
 
 local function cleanUpSessions(sessionNames)
+	if not sessionDataDir then
+		return
+	end
+
 	local validNames = {}
 
 	for _, sessionName in ipairs(sessionNames) do
@@ -524,6 +540,8 @@ local function initialize(event)
 		end
 
 		Observe('PhoneMessagePopupGameController', 'SetTimeDilatation', function(_, popupActive)
+			--spdlog.error(('PhoneMessagePopupGameController::SetTimeDilatation()'))
+
 			updateBlurred(popupActive)
 			notifyObservers()
 		end)
@@ -535,6 +553,8 @@ local function initialize(event)
 
 	if required[GameSession.Scope.Death] and not initialized[GameSession.Scope.Death] then
 		Observe('PlayerPuppet', 'OnDeath', function()
+			--spdlog.error(('PlayerPuppet::OnDeath()'))
+
 			updateDead(true)
 			notifyObservers()
 		end)
@@ -570,7 +590,7 @@ local function initialize(event)
 				table.insert(timestamps, saveInfo.timestamp)
 			end
 
-			dispatchEvent(GameSession.Event.List, { timestamps = timestamps })
+			dispatchEvent(GameSession.Event.Clean, { timestamps = timestamps })
 
 			saveList = nil
 		end)
@@ -621,7 +641,7 @@ local function initialize(event)
 			end
 		end)
 
-		addEventListener(GameSession.Event.List, function(state)
+		addEventListener(GameSession.Event.Clean, function(state)
 			cleanUpSessions(state.timestamps)
 		end)
